backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "backend"
version = "0.1.0"
description = "FastAPI+React full-stack backend template"
keywords=["fastapi","react19","docker/compose","full-stack"]
authors = [
    { name = "Li Ning" ,email="work2lining@163.com"}
]
readme="README.md"
license="MIT"
# 环境
requires-python = ">=3.11"
# 开发环境/核心依赖
dependencies = [
    "fastapi==0.135.0",
    "uvicorn[standard]==0.41.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.0",
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "aiomysql>=0.3.0",
    "pymysql>=1.1.2",
    "redis>=7.2.1",
    "pyjwt>=2.11.0",
    "argon2-cffi>=25.1.0",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "rank-bm25>=0.2.2",
    "loguru>=0.7.3",
    "aiohttp>=3.13.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "mypy>=1.19.1",
    "types-redis>=4.6.0",
    "pre-commit>=4.5.1",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "coverage>=7.13.4",
    "fakeredis>=2.34.1",
    "pytest-xdist>=3.8.0",
    "aiosqlite>=0.22.1",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.hatch.build.targets.wheel]
packages = ["app"]


# Ruff 配置 (代码检查与格式化)
[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "*.pyc",
    "build",
    "dist",
]
[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ASYNC",  # flake8-async
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
]


